Rename setOnClickListener() to init() in LoginActivity
