+++ app/src/main/java/com/avon/spott/EmailLogin/EmailLoginContract.kt
@@ -1,0 +1,2 @@
+package com.avon.spott.EmailLogin
+
